Added .env file creation job in the node.js.yaml file
Because it is impossible to modify serverless file directly in the runner, I am taking the help of dotenv package already in use in the repo to inject environment variables from .env file which will be created through this job in the runner. It will in turn fetch them from repo environment secrets.